+++ DebugFileTools/SymbolsRuntime/runtime_symbols.cpp
@@ -5,7 +5,7 @@
 c_runtime_symbols::c_runtime_symbols(
 	HMODULE module_handle, 
 	void* symbol_file_data, 
-	unsigned long long symbol_file_data_size) :
+	uint64_t symbol_file_data_size) :
 	module_handle(module_handle),
 	symbol_file_header(static_cast<s_symbol_file_header*>(symbol_file_data)),
 	symbol_file_buffer(symbol_file_data),
@@ -34,14 +34,14 @@ void c_runtime_symbols::init()
 	symbol_file_header->binary_name_offset += symbol_file_data_address;
 	symbol_file_header->timestamp_string_offset += symbol_file_data_address;
 
-	for (unsigned long section_index = 0; section_index < symbol_file_header->sections_count; section_index++)
+	for (uint32_t section_index = 0; section_index < symbol_file_header->sections_count; section_index++)
 	{
 		s_symbol_file_section& section = symbol_file_header->sections[section_index];
 		section.class_string_offset += symbol_file_data_address;
 		section.name_string_offset += symbol_file_data_address;
 	}
 
-	for (unsigned long public_symbol_index = 0; public_symbol_index < symbol_file_header->public_symbols_count; public_symbol_index++)
+	for (uint32_t public_symbol_index = 0; public_symbol_index < symbol_file_header->public_symbols_count; public_symbol_index++)
 	{
 		s_symbol_file_public& public_symbol = symbol_file_header->public_symbols[public_symbol_index];
 		public_symbol.lib_and_object_offset += symbol_file_data_address;
@@ -52,7 +52,7 @@ void c_runtime_symbols::init()
 #endif
 	}
 
-	for (unsigned long static_symbol_index = 0; static_symbol_index < symbol_file_header->static_symbols_count; static_symbol_index++)
+	for (uint32_t static_symbol_index = 0; static_symbol_index < symbol_file_header->static_symbols_count; static_symbol_index++)
 	{
 		s_symbol_file_static& static_symbol = symbol_file_header->static_symbols[static_symbol_index];
 		static_symbol.lib_and_object_offset += symbol_file_data_address;
@@ -71,8 +71,8 @@ void c_runtime_symbols::init()
 
 BCS_RESULT c_runtime_symbols::validate_public_symbols()
 {
-	unsigned long long last = 0;
-	for (unsigned long public_symbol_index = 0; public_symbol_index < symbol_file_header->public_symbols_count; public_symbol_index++)
+	uint64_t last = 0;
+	for (uint32_t public_symbol_index = 0; public_symbol_index < symbol_file_header->public_symbols_count; public_symbol_index++)
 	{
 		s_symbol_file_public& public_symbol = symbol_file_header->public_symbols[public_symbol_index];
 
@@ -87,8 +87,8 @@ BCS_RESULT c_runtime_symbols::validate_public_symbols()
 
 BCS_RESULT c_runtime_symbols::validate_static_symbols()
 {
-	unsigned long long last = 0;
-	for (unsigned long static_symbol_index = 0; static_symbol_index < symbol_file_header->static_symbols_count; static_symbol_index++)
+	uint64_t last = 0;
+	for (uint32_t static_symbol_index = 0; static_symbol_index < symbol_file_header->static_symbols_count; static_symbol_index++)
 	{
 		s_symbol_file_static& static_symbol = symbol_file_header->static_symbols[static_symbol_index];
 
@@ -108,7 +108,7 @@ s_symbol_file_public* c_runtime_symbols::get_public_symbol_by_name(const char* s
 		return nullptr;
 	}
 
-	for (unsigned long public_symbol_index = 0; public_symbol_index < symbol_file_header->public_symbols_count; public_symbol_index++)
+	for (uint32_t public_symbol_index = 0; public_symbol_index < symbol_file_header->public_symbols_count; public_symbol_index++)
 	{
 		s_symbol_file_public& public_symbol = symbol_file_header->public_symbols[public_symbol_index];
 
@@ -121,15 +121,15 @@ s_symbol_file_public* c_runtime_symbols::get_public_symbol_by_name(const char* s
 	return nullptr;
 }
 
-s_symbol_file_public* c_runtime_symbols::get_public_symbol_by_relative_virtual_address(unsigned long long relative_virtual_address)
+s_symbol_file_public* c_runtime_symbols::get_public_symbol_by_relative_virtual_address(uint64_t relative_virtual_address)
 {
 	if (symbol_file_header == nullptr)
 	{
 		return nullptr;
 	}
 
-	unsigned long long base_virtual_adress = symbol_file_header->preferred_load_address;
-	unsigned long long relative_virtual_address_plus_base = base_virtual_adress + relative_virtual_address;
+	uint64_t base_virtual_adress = symbol_file_header->preferred_load_address;
+	uint64_t relative_virtual_address_plus_base = base_virtual_adress + relative_virtual_address;
 
 	s_symbol_file_public* public_symbols = symbol_file_header->public_symbols;
 
@@ -143,14 +143,14 @@ s_symbol_file_public* c_runtime_symbols::get_public_symbol_by_relative_virtual_a
 	return public_symbol_search_result;
 }
 
-s_symbol_file_public* c_runtime_symbols::get_public_symbol_by_base_virtual_address(unsigned long long relative_virtual_address)
+s_symbol_file_public* c_runtime_symbols::get_public_symbol_by_base_virtual_address(uint64_t relative_virtual_address)
 {
 	if (symbol_file_header == nullptr)
 	{
 		return nullptr;
 	}
 
-	unsigned long long base_virtual_adress = symbol_file_header->preferred_load_address;
+	uint64_t base_virtual_adress = symbol_file_header->preferred_load_address;
 	return get_public_symbol_by_relative_virtual_address(relative_virtual_address - base_virtual_adress);
 }
 
@@ -161,9 +161,9 @@ s_symbol_file_public* c_runtime_symbols::get_public_symbol_by_virtual_address(co
 		return nullptr;
 	}
 
-	unsigned long long virtual_address = reinterpret_cast<uintptr_t>(pointer);
-	unsigned long long module_address = reinterpret_cast<uintptr_t>(module_handle);
-	unsigned long long relative_virtual_address = virtual_address - module_address;
+	uint64_t virtual_address = reinterpret_cast<uintptr_t>(pointer);
+	uint64_t module_address = reinterpret_cast<uintptr_t>(module_handle);
+	uint64_t relative_virtual_address = virtual_address - module_address;
 
 	return get_public_symbol_by_relative_virtual_address(relative_virtual_address);
 }
